+++ configs/prettier.config.js
@@ -7,4 +7,14 @@ module.exports = {
 	printWidth: 80,
 	arrowParens: 'always',
 	bracketSameLine: false,
+	plugins: ['prettier-plugin-tailwindcss'],
+	overrides: [
+		{
+			files: ['.github/**/*.yml'],
+			options: {
+				useTabs: false,
+				printWidth: 300,
+			},
+		},
+	],
 };