+++ packages/react-kit/.storybook/main.ts
@@ -1,0 +1,21 @@
+import type { StorybookConfig } from '@storybook/react-vite';
+
+const config = {
+	stories: ['../src/**/*.stories.@(js|jsx|mjs|ts|tsx)'],
+	addons: [
+		'@storybook/addon-links',
+		'@storybook/addon-essentials',
+		'@storybook/addon-onboarding',
+		'@storybook/addon-interactions',
+		'@storybook/addon-a11y',
+	],
+	framework: {
+		name: '@storybook/react-vite',
+		options: {},
+	},
+	docs: {
+		autodocs: 'tag',
+	},
+} satisfies StorybookConfig;
+
+export default config;