+++ packages/react-kit/lint-staged.config.cjs
@@ -1,0 +1,3 @@
+const lintStagedConfig = require('@pantheon-systems/workspace-configs/lint-staged');
+
+module.exports = lintStagedConfig;